pyproject.toml
[tool.poetry]
name = "mail-job-tracker"
version = "0.1.0"
description = ""
authors = ["Akash Mishra <akash.mishra13@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.89.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
transformers = "^4.30.2"
torch = "^2.0.1"
openai = "^0.27.8"
pandas = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
